added taskList and parser
