+++ src/data/Yojdata.txt
@@ -1,4 +1,3 @@
-[T] | 0 | ok
-[T] | 1 | trial
-[T] | 0 | lol
-[T] | 0 | haha
+[T] | 1 | ok
+[T] | 0 | trial
+[T] | 1 | lol
