+++ src/main/java/Yoj/Yoj.java
@@ -2,15 +2,16 @@ package Yoj;
 
 import Yoj.exception.InvalidCommandException;
 import Yoj.exception.YojException;
+import Yoj.parser.Parser;
 import Yoj.storage.Storage;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Scanner;
+
+import Yoj.taskList.List;
 import Yoj.ui.*;
 
-import static Yoj.List.List.*;
-import Yoj.List.List.*;
 
 public class Yoj {
 
@@ -18,12 +19,12 @@ public class Yoj {
         Ui.printHello();
         // get user input
         Scanner in = new Scanner(System.in);
-        tasks = Storage.loadTasks();
+        List.tasks = Storage.loadTasks();
         String userInput;
         do {
             userInput = in.nextLine();
             try {
-                manageUserInput(userInput);
+                Parser.manageUserInput(userInput);
             } catch (YojException e){
                 System.out.println(e.getMessage());
             } catch (InvalidCommandException e) {
